+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v0.1.0
+
+beta release
+
+- **breaking**: add / change API functions
+- add mod parameters
+- add `pivoting` mode for setting root note
+- add bend table calculation
+
 ## v0.0.1
 
 alpha release